Add Open Graph image (#158)
* Add ogimage

* change to PNG

og doesn't support svg as image

* Use absolute link

Seems to be a bug of adding additional / to og urls

* Try again

Ok that didn't work either seems baseURL is allowas preended to that. This won't work but want to see what we get

* Fix og image

Found in. was erroneous slash in `url` field that is not handled by og image generation