Update redirects (#187)
* Update redirects

Add redirect rule to catch against previous sites next version

* redirect 1.0.0

Redirect 1.0.0 docs. I hope no one was linking to these!

* Add page specific redirects

In the migration to DSv2 we moved from using the file name in the url to the id. More accurately we defined an id which automatically became the URL. This was needed to ensure the sidebars worked well with the existing nested doc structure. These redirects address the changes so old links should still work. Redirectsto the next or 1.0.0 docs are handled by meeting the wild card rules first and then should be caught by the specific rule for that page (high priority)