Adds 301 redirects to catch more info link from onboarding (#241)
* Remove iOS download links

Temporarily removes the download links for the iOS app while it is not available. Apps notes in appropriate places

* Adds 301 redirects to catch more info link from onboarding

When onboarding failed, some of the error pages contained a link to the error page, however the link is invalid (containing a DSv1 language but not `doc`) and was not caught by existing redirects. 

Here we catch all links of form `/en/*` and send them to `/docs/*` we then also catch `/docs/misc/erros` (previously `/en/misc/errors`) and send that over to `/docs/troubleshooting/errors`