Add notification links (#244)
* Add Android-only icon against clear_notification

It was mentioned in home-assistant/ios#587 that the docs are not clear that `clear_notification` is android only. Added the icon against that line

* Add notification links

You can include a link in the message field so let's tell people about that!

* Mark as iOS only

Currently this method only works on iOS users. Android can use actionable notifications so include a link for them,

* Update docs/notifications/basic.md

Co-authored-by: Daniel Shokouhi <dshokouhi@gmail.com>

Co-authored-by: Daniel Shokouhi <dshokouhi@gmail.com>