Add iOS app link resource
At the moment the app UI has several hard-coded links which can fall out of data easily requiring a new app release. This PR provides a resource to replace the hard coded links.

I've included all the UI links I could spot (quickly) in the source that are for the UI. I have not included auth or API endpoints. And I didn't include the base URLs for Facebook or Twitter because if they change we can presume the service is gone and the links will need removing anyway.

I have updated the companion doc links where needed but have left the What's New link for now as I don't think that is actually done anywhere. I suppose the docs site blog would be the best place?