Add docs index redirect (#253)
In Docusaurus V2 alpha 58 which should drop any moment via dependabot, there is a breaking change that a page will not be built for /docs which we currently have as the getting started guide. This PR just puts in a quick redirect to cover that. Also fixes some errors in the last ones (my bad)