Adds note on Shortcuts that run automatically vs manual
Added which trigger types can work automatically and which ones require manual interaction. Also added one more helpful tip to let you get immediate updates of the battery sensor, same one I shared in Discord earlier today. 

Note: this information applies to iOS 14 and is marked as such. I think the target audience for Shortcuts (advanced users) are the same ones that jump on the dev/public beta train, and for everyone else the final stable release is only like ~2 months away. But if we'd rather hold off on publishing this until it officially releases, that is fine with me too. Will leave that call up to you Tom.