clarification for command_launch_app
Each time i try to use this feature, I forget how to find the "package_name". Hopefully this additional sentence in the doc will help me and others in the future.
This was a relatively easy google on my part... but should we really expect people to google to find how to use "command_launch_app"?